ref(StoreAuth):  modified login & logout
remove setting a token on login to prevent duplication from the `getUser()`, modified catch block on the login that removes the token if the any errors occur.  And refactor logout that uses the synchronous `getItem` from the local storage instead of using the state of token which has a value of null from init.